Update control-components
- Add support for analog soil sensor configuration via HTTP API
- Add support for analog LDR sensor configuration via HTTP API
